+++ render/extension.ts
@@ -5,7 +5,8 @@ import Parser from '../index';
 import type {ExtToken, GalleryToken, Token, NowikiToken} from '../internal';
 
 export const packedModes = new Set<string | undefined>(['packed', 'packed-hover', 'packed-overlay']);
-const galleryModes = new Set([...packedModes, 'nolines', 'slideshow']);
+const galleryModes = new Set([...packedModes, 'nolines', 'slideshow']),
+	templateDataSortValues = {deprecated: -1, suggested: 1, required: 2, optional: 0};
 
 /** @ignore */
 const getCiteNoteId = (i: number, refName?: string): string =>
@@ -346,6 +347,92 @@ export const renderExt = (token: ExtToken, opt?: Omit<HtmlOpt, 'nowrap'>): strin
 			}"${id ? ` id="${sanitizeId(id)}"` : ''}>${
 				texToSvg?.((lastChild as NowikiToken).texvcCheck()!) ?? ''
 			}</span>`;
+		}
+		case 'templatedata': {
+			const {
+				validateTemplateData,
+				resolveInterfaceText,
+				resolveOptionalInterfaceText,
+			}: typeof import('./templatedata') = require('./templatedata');
+			const data = validateTemplateData(token.innerText ?? '');
+			if (typeof data === 'string') {
+				return '<div class="cdx-message--error cdx-message cdx-message--block>'
+					+ '<span class="cdx-message__icon"></span>'
+					+ `<div class="cdx-message__content"><p>${sanitize(data)}\n</p></div>`
+					+ '</div>';
+			}
+			const {description, params, paramOrder = Object.keys(params)} = data;
+			return '<section class="mw-templatedata-doc-wrap">'
+				+ `<header><p class="mw-templatedata-doc-desc${
+					description === undefined ? ' mw-templatedata-doc-muted' : ''
+				}">${description === undefined ? 'No description.' : resolveInterfaceText(description)}</p></header>`
+				+ `<table class="wikitable mw-templatedata-doc-params${
+					paramOrder.length === 0 ? '' : ' sortable'
+				}">`
+				+ '<caption><p class="mw-templatedata-caption">Template parameters</p></caption>'
+				+ '<thead>'
+				+ '<tr><th colspan="2">Parameter</th><th>Description</th><th>Type</th><th>Status</th></tr>'
+				+ '</thead>'
+				+ `<tbody>${
+					paramOrder.length === 0
+						? '<tr><td class="mw-templatedata-doc-muted" colspan="7">No parameters specified</td></tr>'
+						: paramOrder.map(p => {
+							const {
+								aliases = [],
+								suggestedvalues,
+								deprecated,
+								required,
+								suggested,
+								label = p,
+								description: desc,
+								default: d,
+								example,
+								autovalue,
+								type = 'unknown',
+							} = params[p]!;
+							let status: 'deprecated' | 'optional' | 'required' | 'suggested' = 'optional';
+							if (deprecated) {
+								status = 'deprecated';
+							} else if (required) {
+								status = 'required';
+							} else if (suggested) {
+								status = 'suggested';
+							}
+							return '<tr>'
+								+ `<th>${resolveInterfaceText(label)}</th>`
+								+ `<td class="mw-templatedata-doc-param-name">${
+									[
+										`<code>${sanitize(p)}</code>`,
+										...aliases.map(alias => `<code class="mw-templatedata-doc-param-alias">${
+											sanitize(alias)
+										}</code>`),
+									].join(' ')
+								}</td>`
+								+ '<td>'
+								+ `<p${desc ? '' : ' class="mw-templatedata-doc-muted"'}>${
+									desc === undefined ? 'no description' : resolveInterfaceText(desc)
+								}</p>`
+								+ `<dl>${
+									suggestedvalues
+										? `<dt>Suggested values</dt><dd>${
+											suggestedvalues.map(v => `<code>${sanitize(v)}</code>`).join(' ')
+										}</dd>`
+										: ''
+								}${resolveOptionalInterfaceText('Default', d)}${
+									resolveOptionalInterfaceText('Example', example)
+								}${resolveOptionalInterfaceText('Auto value', autovalue, true)}</dl>`
+								+ '</td>'
+								+ `<td class="mw-templatedata-doc-param-type${
+									type === 'unknown' ? ' mw-templatedata-doc-muted' : ''
+								}">${type}</td>`
+								+ `<td class="mw-templatedata-doc-param-status-${status}" data-sort-value="${
+									templateDataSortValues[status]
+								}">${status}</td>`
+								+ '</tr>';
+						}).join('')
+				}</tbody>`
+				+ '</table>'
+				+ '</section>';
 		}
 		default:
 			return '';